build(setup):Eslint/Prettier/Husky
