+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+  "useTabs": true,
+  "tabSize": 2,
+  "semi": true,
+  "singleQuote": true,
+  "trailingComma": "es5",
+  "bracketSpacing": true,
+  "parser": "typescript"
+}
