+++ 6
@@ -1,0 +1,32 @@
+yarn add v1.22.11
+[1/4] Resolving packages...
+[2/4] Fetching packages...
+info fsevents@2.3.2: The platform "win32" is incompatible with this module.
+info "fsevents@2.3.2" is an optional dependency and failed compatibility check. Excluding it from installation.
+info fsevents@1.2.13: The platform "win32" is incompatible with this module.
+info "fsevents@1.2.13" is an optional dependency and failed compatibility check. Excluding it from installation.
+[3/4] Linking dependencies...
+[4/4] Building fresh packages...
+success Saved lockfile.
+success Saved 14 new dependencies.
+info Direct dependencies
+├─ husky@7.0.2
+└─ lint-staged@11.1.2
+info All dependencies
+├─ cli-cursor@3.1.0
+├─ human-signals@2.1.0
+├─ husky@7.0.2
+├─ is-unicode-supported@0.1.0
+├─ lint-staged@11.1.2
+├─ listr2@3.11.1
+├─ log-symbols@4.1.0
+├─ log-update@4.0.0
+├─ please-upgrade-node@3.2.0
+├─ restore-cursor@3.1.0
+├─ rxjs@6.6.7
+├─ semver-compare@1.0.0
+├─ string-argv@0.3.1
+└─ through@2.3.8
+$ husky install
+husky - Git hooks installed
+Done in 141.49s.
