+++ src/App.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <h1>hey</h1>
+    </div>
+  );
+}
+
+export default App;
